[MRC] Explorer: see an overview of all projects I'm contributing to on cart checkout page (#2101)
