+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 /lerna-debug.log
 .turbo
 .husky/
+.vercel
