+++ packages/grant-explorer/package.json
@@ -35,17 +35,18 @@
     ]
   },
   "dependencies": {
+    "@chakra-ui/react": "^2.1.2",
     "@craco/craco": "^7.0.0",
     "@datadog/browser-logs": "^4.19.0",
     "@datadog/browser-rum": "^4.19.0",
+    "@emotion/react": "^11",
+    "@emotion/styled": "^11",
     "@ethersproject/abstract-signer": "^5.7.0",
     "@ethersproject/providers": "^5.7.2",
     "@gitcoinco/passport-sdk-verifier": "^0.2.2",
     "@headlessui/react": "^1.7.4",
     "@heroicons/react": "^2.0.13",
-    "@lagunovsky/redux-react-router": "^3.1.0",
     "@rainbow-me/rainbowkit": "0.12.16",
-    "@reduxjs/toolkit": "^1.8.1",
     "@sentry/integrations": "^7.28.0",
     "@sentry/react": "^7.27.0",
     "@sentry/tracing": "^7.26.0",
@@ -68,6 +69,7 @@
     "eslint": ">=8.1.0 <9.0.0",
     "ethereum-blockies": "^0.1.1",
     "ethers": "^5.7.2",
+    "framer-motion": "^10.15.0",
     "history": "^5.3.0",
     "html-react-parser": "^3.0.7",
     "https-browserify": "^1.0.0",
@@ -85,23 +87,21 @@
     "react-dom": "^18.1.0",
     "react-gtm-module": "^2.0.11",
     "react-hook-form": "^7.31.3",
-    "react-redux": "^8.0.1",
     "react-router": "^6",
     "react-router-dom": "6",
     "react-scripts": "5.0.1",
     "react-tooltip": "^4.5.0",
-    "redux": "latest",
-    "redux-thunk": "latest",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "swr": "^2.0.0",
-    "typescript": "^4.6.0",
+    "typescript": "^5.1.6",
     "url": "^0.11.0",
     "util": "^0.12.4",
     "viem": "^1.4.2",
     "wagmi": "0.12.18",
     "web-vitals": "^2.1.0",
-    "webpack": ">=2"
+    "webpack": ">=2",
+    "zustand": "^4.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.12",
