+++ packages/grant-explorer/src/features/api/rounds.ts
@@ -112,17 +112,10 @@ function getActiveChainIds() {
 
 export async function getRoundsInApplicationPhase(
   debugModeEnabled: boolean
-): Promise<{
-  isLoading: boolean;
-  error: unknown;
-  rounds: RoundOverview[];
-}> {
-  try {
-    const chainIds = getActiveChainIds();
+): Promise<RoundOverview[]> {
+  const chainIds = getActiveChainIds();
 
-    const rounds: RoundOverview[] = [];
-
-    const query = `
+  const query = `
       query GetRoundsInApplicationPhase($currentTimestamp: String) {
         rounds(where: {
           applicationsStartTime_lt: $currentTimestamp
@@ -149,41 +142,21 @@ export async function getRoundsInApplicationPhase(
       }
     `;
 
-    for (let i = 0; i < chainIds.length; i++) {
-      const roundsForChainId = await fetchRoundsByTimestamp(
-        query,
-        chainIds[i],
-        debugModeEnabled
-      );
-      rounds.push(...roundsForChainId);
-    }
+  const rounds = await Promise.all(
+    chainIds.map((chainId) =>
+      fetchRoundsByTimestamp(query, chainId, debugModeEnabled)
+    )
+  );
 
-    return {
-      isLoading: false,
-      error: undefined,
-      rounds,
-    };
-  } catch (error) {
-    console.error("getRoundsInApplicationPhase", error);
-    return {
-      isLoading: false,
-      error,
-      rounds: [],
-    };
-  }
+  return rounds.flat();
 }
 
-export async function getActiveRounds(debugModeEnabled: boolean): Promise<{
-  isLoading: boolean;
-  error: unknown;
-  rounds: RoundOverview[];
-}> {
-  try {
-    let isLoading = true;
-    const chainIds = getActiveChainIds();
-    const rounds: RoundOverview[] = [];
+export async function getActiveRounds(
+  debugModeEnabled: boolean
+): Promise<RoundOverview[]> {
+  const chainIds = getActiveChainIds();
 
-    const query = `
+  const query = `
       query GetActiveRounds($currentTimestamp: String) {
         rounds(where: {
           roundStartTime_lt: $currentTimestamp
@@ -210,30 +183,11 @@ export async function getActiveRounds(debugModeEnabled: boolean): Promise<{
       }
     `;
 
-    for (let i = 0; i < chainIds.length; i++) {
-      console.log(chainIds);
-      const roundsForChainId = await fetchRoundsByTimestamp(
-        query,
-        chainIds[i],
-        debugModeEnabled
-      );
-      rounds.push(...roundsForChainId);
-    }
+  const rounds = await Promise.all(
+    chainIds.map((chainId) =>
+      fetchRoundsByTimestamp(query, chainId, debugModeEnabled)
+    )
+  );
 
-    const error = undefined;
-    isLoading = false;
-
-    return {
-      isLoading,
-      error,
-      rounds,
-    };
-  } catch (error) {
-    console.error("getActiveRounds", error);
-    return {
-      isLoading: false,
-      error,
-      rounds: [],
-    };
-  }
+  return rounds.flat();
 }
