--- packages/grant-explorer/src/features/api/services/ipfs.ts
@@ -1,42 +1,0 @@
-import { IPFSObject } from "../types";
-import { fetchFromIPFS, pinToIPFS } from "../utils";
-import { api } from "..";
-
-export const ipfsApi = api.injectEndpoints({
-  endpoints: (builder) => ({
-    saveToIPFS: builder.mutation<string, IPFSObject>({
-      queryFn: async (object) => {
-        const result = { data: "" };
-
-        try {
-          const resp = await pinToIPFS(object);
-
-          console.log("Added file to IPFS:", resp.IpfsHash);
-          result.data = resp.IpfsHash;
-
-          return result;
-        } catch (err) {
-          console.error("saveToIPFS", err);
-          return { error: "Unable to save file to IPFS" };
-        }
-      },
-      invalidatesTags: ["IPFS"],
-    }),
-    readFromIPFS: builder.query<string, string>({
-      queryFn: async (cid) => {
-        try {
-          const data = await fetchFromIPFS(cid);
-
-          return { data };
-        } catch (err) {
-          console.log("readFromIPFS", err);
-          return { error: "Unable to fetch file from IPFS" };
-        }
-      },
-      providesTags: ["IPFS"],
-    }),
-  }),
-  overrideExisting: false,
-});
-
-export const { useSaveToIPFSMutation, useReadFromIPFSQuery } = ipfsApi;
