+++ packages/grant-explorer/src/features/common/NavbarCart.tsx
@@ -1,26 +1,17 @@
-import { Link, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { Project } from "../api/types";
-import { Button } from "common/src/styles";
-import { CheckIcon, EyeIcon } from "@heroicons/react/24/outline";
-import DefaultLogoImage from "../../assets/default_logo.png";
-import { renderToPlainText } from "common";
 import tw from "tailwind-styled-components";
 
-export default function NavbarCart(props: {
-  cart: Project[];
-  roundUrlPath: string;
-}) {
+export default function NavbarCart(props: { cart: Project[] }) {
   const projectCount = props.cart.length;
 
   const navigate = useNavigate();
-
-  // eslint-disable-next-line react-hooks/exhaustive-deps
 
   return (
     <div
       data-testid="navbar-cart"
       className="relative flex-row"
-      onClick={() => navigate(`${props.roundUrlPath}/cart`)}
+      onClick={() => navigate(`/cart`)}
     >
       <QuickViewIcon count={projectCount} />
     </div>
@@ -71,99 +62,6 @@ function QuickViewIcon(props: { count: number }) {
           {props.count}
         </Badge>
       )}
-    </div>
-  );
-}
-
-export function QuickViewSummary(props: {
-  roundUrlPath: string;
-  cart: Project[];
-}) {
-  const navigate = useNavigate();
-
-  const QuickViewSummary = tw.div`
-    mt-5
-    w-[400px]
-    flex
-    flex-col
-    absolute
-    right-0
-    rounded
-    bg-white
-    shadow-[0px_0px_10px_0px_rgba(0,0,0,0.1)]
-    p-4
-    z-20
-  `;
-
-  return (
-    <QuickViewSummary
-      data-testid="quick-view-summary"
-      className="quick-view-summary"
-    >
-      <p className="flex border-b mb-4 pb-2 gap-2 text-sm">
-        <CheckIcon className="w-4" />
-        Projects added to your cart
-      </p>
-
-      <div className="max-h-[200px] overflow-y-scroll">
-        {props.cart &&
-          props.cart.map((project) => (
-            <ProjectQuickView
-              roundUrlPath={props.roundUrlPath}
-              project={project}
-            />
-          ))}
-      </div>
-
-      <Button
-        type="button"
-        $variant="solid"
-        className="px-3 bg-violet-400 text-white border-0 text-xs mb-2"
-        onClick={() => navigate(`${props.roundUrlPath}/cart`)}
-      >
-        View my cart ({props.cart.length})
-      </Button>
-    </QuickViewSummary>
-  );
-}
-
-function ProjectQuickView(props: { roundUrlPath: string; project: Project }) {
-  return (
-    <div className="flex mb-4" data-testid="project-quick-view">
-      <div className="relative overflow-hidden bg-no-repeat bg-cover  min-w-[64px] w-16 max-h-[64px] mt-auto mb-auto">
-        <img
-          className="inline-block"
-          src={
-            props.project.projectMetadata.logoImg
-              ? `https://${process.env.REACT_APP_PINATA_GATEWAY}/ipfs/${props.project.projectMetadata.logoImg}`
-              : DefaultLogoImage
-          }
-          alt={"Project Logo"}
-        />
-        <div className="min-w-[64px] w-16 max-h-[64px] absolute top-0 right-0 bottom-0 left-0 overflow-hidden bg-fixed opacity-0 hover:opacity-70 transition duration-300 ease-in-out bg-gray-500 justify-center flex items-center">
-          <Link
-            to={`${props.roundUrlPath}/${props.project.grantApplicationId}`}
-          >
-            <EyeIcon
-              className="fill-gray-200 w-6 h-6 cursor-pointer"
-              data-testid={`${props.project.projectRegistryId}-project-link`}
-            />
-          </Link>
-        </div>
-      </div>
-
-      <div className="pl-4 mt-1">
-        <Link to={`${props.roundUrlPath}/${props.project.grantApplicationId}`}>
-          <p className="font-semibold mb-2 text-sm text-ellipsis line-clamp-1">
-            {props.project.projectMetadata.title}
-          </p>
-        </Link>
-        <p className="text-xs text-ellipsis line-clamp-3">
-          {renderToPlainText(
-            props.project.projectMetadata.description
-          ).substring(0, 20)}
-        </p>
-      </div>
     </div>
   );
 }
