+++ packages/grant-explorer/src/features/discovery/ActiveRoundSection.tsx
@@ -10,12 +10,12 @@ import RoundCard from "./RoundCard";
 type ActiveRounds = {
   isLoading: boolean;
   setSearchQuery: Dispatch<SetStateAction<string>>;
-  roundOverview: RoundOverview[];
+  roundOverview?: RoundOverview[];
   searchQuery: string;
 };
 
 const ActiveRoundsSection = (props: ActiveRounds) => {
-  const activeRoundsCount = props.roundOverview.length;
+  const activeRoundsCount = props.roundOverview?.length ?? 0;
   const [order, setOrder] = useState<string>("round_asc");
 
   function sortRoundsByTime(rounds: RoundOverview[], order: string) {
@@ -71,7 +71,7 @@ const ActiveRoundsSection = (props: ActiveRounds) => {
         ) : null}
         {activeRoundsCount > 0 ? (
           <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 md:gap-6 2xl:grid-cols-4">
-            {sortRoundsByTime(props.roundOverview, order)
+            {sortRoundsByTime(props.roundOverview ?? [], order)
               .filter((round) => (round.projects?.length ?? 0) > 0)
               // .filter((round) => round.chainId === "PGN_TESTNET")
               .map((round, index) => {
