+++ packages/grant-explorer/src/index.tsx
@@ -1,26 +1,26 @@
 import "./browserPatches";
 
-import { ReduxRouter } from "@lagunovsky/redux-react-router";
 import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { Provider } from "react-redux";
-import { Route, Routes } from "react-router-dom";
+import {
+  BrowserRouter,
+  HashRouter,
+  Route,
+  Router,
+  Routes,
+} from "react-router-dom";
 import { WagmiConfig } from "wagmi";
 import { RoundProvider } from "./context/RoundContext";
 import { initDatadog } from "./datadog";
 import { initSentry } from "./sentry";
 import { initTagmanager } from "./tagmanager";
-
-import { store } from "./app/store";
 import { chains, client as WagmiClient } from "./app/wagmi";
-import history from "./history";
 import reportWebVitals from "./reportWebVitals";
 
 import "./index.css";
 
 // Routes
-import { CartProvider } from "./context/CartContext";
 import { QFDonationProvider } from "./context/QFDonationContext";
 import AccessDenied from "./features/common/AccessDenied";
 import Auth from "./features/common/Auth";
@@ -29,10 +29,12 @@ import ApplyNowPage from "./features/discovery/ApplyNowPage";
 import LandingPage from "./features/discovery/LandingPage";
 import PassportConnect from "./features/round/PassportConnect";
 import ThankYou from "./features/round/ThankYou";
-import ViewCart from "./features/round/ViewCartPage";
 import ViewProjectDetails from "./features/round/ViewProjectDetails";
 import ViewRound from "./features/round/ViewRoundPage";
 import ViewContributionHistory from "./features/contributors/ViewContributionHistory";
+import ViewCart from "./features/round/ViewCartPage/ViewCartPage";
+import { Switch } from "@headlessui/react";
+import { ChakraProvider } from "@chakra-ui/react";
 
 // Initialize sentry
 initSentry();
@@ -49,73 +51,71 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <Provider store={store}>
+    <ChakraProvider>
       <WagmiConfig client={WagmiClient}>
         <RainbowKitProvider coolMode chains={chains}>
           <RoundProvider>
-            <CartProvider>
-              <ReduxRouter history={history} store={store}>
-                <Routes>
-                  {/* Protected Routes */}
-                  <Route element={<Auth />} />
+            <HashRouter>
+              <Routes>
+                {/* Protected Routes */}
+                <Route element={<Auth />} />
 
-                  {/* Default Route */}
-                  <Route path="/" element={<LandingPage />} />
+                {/* Default Route */}
+                <Route path="/" element={<LandingPage />} />
 
-                  {/* Apply Now Page */}
-                  <Route path="/apply-now" element={<ApplyNowPage />} />
+                {/* Apply Now Page */}
+                <Route path="/apply-now" element={<ApplyNowPage />} />
 
-                  {/* Round Routes */}
-                  <Route
-                    path="/round/:chainId/:roundId"
-                    element={<ViewRound />}
-                  />
-                  <Route
-                    path="/round/:chainId/:roundId/:applicationId"
-                    element={<ViewProjectDetails />}
-                  />
+                {/* Round Routes */}
+                <Route
+                  path="/round/:chainId/:roundId"
+                  element={<ViewRound />}
+                />
+                <Route
+                  path="/round/:chainId/:roundId/:applicationId"
+                  element={<ViewProjectDetails />}
+                />
 
-                  <Route
-                    path="/round/:chainId/:roundId/cart"
-                    element={
-                      <QFDonationProvider>
-                        <ViewCart />
-                      </QFDonationProvider>
-                    }
-                  />
+                <Route
+                  path="/cart"
+                  element={
+                    <QFDonationProvider>
+                      <ViewCart />
+                    </QFDonationProvider>
+                  }
+                />
 
-                  <Route
-                    path="/round/:chainId/:roundId/:txHash/thankyou"
-                    element={
-                      <QFDonationProvider>
-                        <ThankYou />
-                      </QFDonationProvider>
-                    }
-                  />
+                <Route
+                  path="/thankyou"
+                  element={
+                    <QFDonationProvider>
+                      <ThankYou />
+                    </QFDonationProvider>
+                  }
+                />
 
-                  {/* Passport Connect */}
-                  <Route
-                    path="/round/:chainId/:roundId/passport/connect"
-                    element={<PassportConnect />}
-                  />
+                {/* Passport Connect */}
+                <Route
+                  path="/round/:chainId/:roundId/passport/connect"
+                  element={<PassportConnect />}
+                />
 
-                  <Route
-                    path="/contributors/:address"
-                    element={<ViewContributionHistory />}
-                  />
+                <Route
+                  path="/contributors/:address"
+                  element={<ViewContributionHistory />}
+                />
 
-                  {/* Access Denied */}
-                  <Route path="/access-denied" element={<AccessDenied />} />
+                {/* Access Denied */}
+                <Route path="/access-denied" element={<AccessDenied />} />
 
-                  {/* 404 */}
-                  <Route path="*" element={<NotFound />} />
-                </Routes>
-              </ReduxRouter>
-            </CartProvider>
+                {/* 404 */}
+                <Route path="*" element={<NotFound />} />
+              </Routes>
+            </HashRouter>
           </RoundProvider>
         </RainbowKitProvider>
       </WagmiConfig>
-    </Provider>
+    </ChakraProvider>
   </React.StrictMode>
 );
